requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
av>=10.0.0
# mmcv>=2.0.0
# mmpose>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0
langchain>=0.0.300
chromadb>=0.4.0
sentence-transformers>=2.2.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.65.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
